fix: redux example
