+++ angular-universal-ssr/host-app/webpack/host-ssr.js
@@ -49,7 +49,7 @@ module.exports = (env = {}) => {
     ],
 
     module: {
-      rules: [ ...require('./_loaders') ]
-    }
+      rules: [...require("./_loaders")],
+    },
   };
 };
