+++ nested/app1/src/index.js
@@ -1,5 +1,2 @@
-import App from "./App";
-import React from "react";
-import ReactDOM from "react-dom";
-
-ReactDOM.render(<App />, document.getElementById("root"));
+// theres a better way to do this. Im just lazy right now
+import("./bootstrap");
