+++ nested/app3/src/index.js
@@ -1,5 +1,1 @@
-import App from "./App";
-import React from "react";
-import ReactDOM from "react-dom";
-
-ReactDOM.render(<App />, document.getElementById("root"));
+import("./bootstrap");
