+++ nested/app3/webpack.config.js
@@ -31,7 +31,7 @@ module.exports = {
       library: { type: "var", name: "app3" },
       filename: "remoteEntry.js",
       exposes: {
-        Button: "./src/Button",
+        "./Button": "./src/Button",
       },
       shared: ["react", "react-dom"],
     }),
